Src/requirement.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==20.3.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==20.4.1
configobj==5.0.6
cryptography==3.3.2
exceptiongroup==1.2.0
fastapi==0.109.2
geojson==3.1.0
h11==0.14.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
iotop==0.6
Jinja2==2.11.3
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==4.2.0
numpy==1.26.4
oauthlib==3.1.0
psycopg2-binary==2.9.9
pycurl==7.43.0.6
pydantic==2.6.1
pydantic-core==2.16.2
PyJWT==1.7.1
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
PyYAML==5.3.1
reportbug==7.10.3+deb11u1
requests==2.25.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing-extensions==4.9.0
unicorn==2.0.1.post1
urllib3==1.26.5
uvicorn==0.27.0.post1
zipp==1.0.0
